SEQUOIADBMAINSTREAM-6563: fix arm64 align issue

+++ SequoiaDB/engine/include/mthMatchNode.hpp
@@ -375,7 +375,7 @@ namespace engine
    // 2. By using user specified allocator(instances of _mthNodeAllocator).
    // The actions when releasing these two kinds of node are different.
    // So a type flag is added at the head of the actual allocated memory.
-   #define MTH_MEM_TYPE_SIZE            sizeof(INT32)
+   #define MTH_MEM_TYPE_SIZE            sizeof(INT64)
    #define MTH_MEM_BY_USER_ALLOCATOR    0
    #define MTH_MEM_BY_DFT_ALLOCATOR     1
